refactor: Board 관련 VO domain 엔티티로 패키지 위치 변경
