+++ src/main/java/com/example/demo/application/board/dto/request/PresignedUrlRequest.java
@@ -1,6 +1,6 @@
 package com.example.demo.application.board.dto.request;
 
-import com.example.demo.application.board.dto.vo.FileType;
+import com.example.demo.domain.board.service.entity.vo.FileType;
 import com.example.demo.global.aop.valid.ValidEnum;
 
 import io.swagger.v3.oas.annotations.media.Schema;
