+++ src/main/java/com/example/demo/domain/board/service/implement/BoardReader.java
@@ -9,8 +9,7 @@ import com.example.demo.infra.board.Repository.BoardJpaRepository;
 import com.example.demo.application.board.dto.request.BoardUpdateRequest;
 import com.example.demo.domain.board.service.entity.BoardTitleInfo;
 import com.example.demo.domain.board.service.entity.DraftBoardTitle;
-import com.example.demo.application.board.dto.vo.BoardType;
-import com.example.demo.global.base.dto.page.GlobalPageResponse;
+import com.example.demo.domain.board.service.entity.vo.BoardType;
 import com.example.demo.infra.board.entity.Board;
 import com.example.demo.global.base.exception.ErrorCode;
 import com.example.demo.global.base.exception.ServiceException;
