+++ src/main/java/com/example/demo/global/base/exception/ErrorCode.java
@@ -1,7 +1,7 @@
 package com.example.demo.global.base.exception;
 
-import com.example.demo.application.board.dto.vo.BoardType;
-import com.example.demo.application.board.dto.vo.Status;
+import com.example.demo.domain.board.service.entity.vo.BoardType;
+import com.example.demo.domain.board.service.entity.vo.Status;
 import com.example.demo.domain.comment.domain.vo.CommentTargetBoardType;
 import com.example.demo.domain.recruitment_board.domain.vo.EntireBoardType;
 import com.example.demo.domain.recruitment_board.domain.vo.QuestionType;
