+++ src/main/java/com/example/demo/global/converter/StatusConverter.java
@@ -1,6 +1,6 @@
 package com.example.demo.global.converter;
 
-import com.example.demo.application.board.dto.vo.Status;
+import com.example.demo.domain.board.service.entity.vo.Status;
 import com.example.demo.global.base.exception.ErrorCode;
 import com.example.demo.global.base.exception.ServiceException;
 import org.springframework.core.convert.converter.Converter;
