+++ src/main/java/com/example/demo/infra/board/querydsl/BoardDslRepository.java
@@ -5,8 +5,7 @@ import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import com.example.demo.application.board.dto.vo.BoardType;
-import com.example.demo.domain.base.page.GlobalPageableDto;
+import com.example.demo.domain.board.service.entity.vo.BoardType;
 import com.example.demo.domain.board.service.entity.BoardTitleInfo;
 import com.example.demo.domain.board.service.entity.DraftBoardTitle;
 import com.example.demo.infra.board.entity.Board;
