+++ src/test/java/com/example/demo/infra/fixture/board/BoardFixtures.java
@@ -1,7 +1,7 @@
 package com.example.demo.infra.fixture.board;
 
-import com.example.demo.application.board.dto.vo.BoardType;
-import com.example.demo.application.board.dto.vo.Status;
+import com.example.demo.domain.board.service.entity.vo.BoardType;
+import com.example.demo.domain.board.service.entity.vo.Status;
 import com.example.demo.infra.board.entity.Board;
 import com.example.demo.domain.user.domain.User;
